uint16_t in Robot (correction)

+++ src/robot.h
@@ -42,9 +42,9 @@ public:
     double ETA();
 
     // 外部决定属性
-    uint8_t id;                // 机器人ID
+    uint16_t id;                // 机器人ID
     int workbench_id;          // 所处工作台id
-    uint8_t item_type;         // 携带物品类型
+    uint16_t item_type;         // 携带物品类型
     double time_val = 0.0;     // 时间价值系数
     double collision_val = 0.0;// 碰撞价值系数
     double omega = 0.0;        // 角速度
